fix: prevent a help lookup crash
This code uses DescMenu, which for some reason, is a subclass of
InvMenu, which is expecting its items to correspond to inventory items.
We need to override on_examine therefore to prevent crashes. The
following comment in directn.cc unfortunately still remains unresolved:

`// XX this probably shouldn't use InvMenu, why does it?`