feat: default `reduce_animations` to true on webtiles
Crawl's animations have long been a problem for webtiles. They work by
using precisely timed delays (ranging from as short as 25ms up to a few
hundred ms) combined with many redraws at interim stages as e.g. a beam
path unfolds square by square. This works fine locally, but on webtiles
this is all run on the server side, so a redraw is (a) at the mercy of
the network, which can impact this timing substantially, and (b) causes
a spike in map data sent to the server. Ideally, if animations are
desirable on webtiles it would be much, much better to run them in the
client, but this is something noone has ever implemented (and it doesn't
sound very easy to do). The situation is so bad that stronger players
tend to just turn off animations altogether. However, doing so leads to
no visual feedback at all for ranged attacks that use animations.  This
commit threads the needle a bit, and by default shows most animations
with just one delay + one graphic. It's possible this will still be too
underinformative for newer players, and this commit might need to be
reverted.

As more background: A worst case scenario has been that firing or
drawing beam animations became so laggy on webtiles, that in some
conditions it has been slower than the key repeat timing or are slow
enough for lag to impact webtiles server performance, see e.g.
0b5ff556b1e2, 0306a414dbaaa, e645bd13c135ed, etc. These commits involved
hotspots that were improvable, but I think this option potentially
provides a much better default solution.