feat: better webtiles username ban management
Since 0.28 or so, I have behind the scenes been distributing code to
server admins that bans some slurs in usernames for account creation.
This commit substantially enhances that code and builds it in to the
webtiles server. This does not add any banned usernames to the repo /
default config, but does make them quite a bit easier to add.

Highlights:
1. A config option `banned` that can be used to indicate disallowed
   usernames, including handling some basic leet-speak translation,
   letter repeats, etc. Currently, these are disallowed for account
   creation only, and database bans are a distinct concept (but see 4
   below). See documentation in config.py on how to use the more
   complex possibilities here.
2. The ability to load this list from a yml (or txt) file, similar to
   Alex's config.yml override. The format of the option is designed to
   be easily encoded in yml, and also line up with the (simpler) yml
   banned names format for the scoring db. This also means that admins
   no longer have to see / have visible on their screen slurs that are
   written out in a config file. (To be honest, this problem was the
   main impetus for me starting work on this set of changes.)
3. A distinct bans.py module that is self-contained and can be pulled
   into other crawl projects (e.g. scoring, tourney). I'm hoping that we
   can start to share some bits of infrastructure-wide config like this
   across servers, projects, etc...
4. Options for the command line utility that can check config-based bans
   against the user database, and automatically ban accounts.

As noted in several comments, this code makes it very easy for admins to
run into https://en.wikipedia.org/wiki/Scunthorpe_problem, so exercise
care with the more powerful username matching options.  There's
preliminary allowlist code that can be expanded if this is desireable.

This probably makes it overall easier for an attacker to come up with
ways around username restrictions (by removing some things, such as my
exact leet-speak decoding, from obscurity). But, I do hope that this
commit still overall improves the situation for server admins. In the
worst case scenario this is no different than the the previous state of
affairs: it acts as a preventative for more casual abuse, and account
holds fill the role of dealing with more persistent attackers.