Fix Ru faith penalties always applying to Reject Sacrifices
Since 26bd780f50c, manually rejecting a set of sacrifices via the Ru
ability has applied a bonus penalty, which was intended to apply only when
an amulet of faith was removed. Fixing this is a reasonably large buff to
rejecting sacrifices (especially early ones), but hopefully that's fine
considering that rejecting sacrifices is usually a bad idea anyway. If not
then the base penalties could easily be increased.