feat: show webtiles socket info to admin users
On a reload, show current socket stats and username info in the admin
panel.  This does not refresh except manually.  This may or may not turn
out to be useful, but otherwise there's right now no good way of getting
info about usernames in the lobby or spectating. This also lets admin
users more easily see what held accounts are currently doing.