Adjust noise for lightning-based spells
Lightning Bolt, Arcjolt and Chain Lightning are all intended to be
especially noisy, but the changes to zap noise now cause them to create
their noise on every target hit, instead of just once at the casting
position. Rather than special-casing them again, keep the new behaviour but
reduce the bonus noise for Lightning Bolt and Arcjolt. Chain Lightning gets
to just be extremely noisy, as is fitting for a level 9 spell.