Adding a trireme
I've wanted to add something that looks like a ship to shoals, here we go! This one is packed full of lots of nasty humans, should make for some interesting challenges to clear!

Feel free to change the balance of monsters in this if required. I'm also certainly willing to consider alternative materials. I toyed with crystal or metal, but I think this will work fine as is.