Let monster zaps make noise on hit
When setting up beams for zap-based monster spells, the loudness value was
discarded, making them silent on hit. There's no obvious reason that they
should work differently, so allow them to work the same way as player zaps,
and make noise happen when the player is hit by a beam.