feat: full CLI for dgl flag management
Previously, for anything other than bans or holds you needed to use
dgamelaunch-config scripts, or set the flags manually via sqlite itself.
This provides a cli for setting all (in-use) flags. It also updates to
reserve the new flags in use on CDI; these are not currently used by
webtiles at all except via this interface, and there's some cosmetic
options that alert the user about this.

This commit also refactors some of the other CLI subcommands and
improves error messages.