fix: autotargeting for Vhi's electric charge
This was quite involved because Vhi's has its own special targeting code
(from when it was an ability). This code is involved enough that I don't
immediately know exactly how to merge it with more general spell
targeting code. This commit at least removes some of the special casing,
but not all by a long stretch -- just enough to support non-interactive
targeting (e.g. autofiring via the quiver).