feat: enhance templating for webtiles game configs
Previously, these allowed using the template string "%n" in certain
game config fields, which would be replaced by the username. This commit
expands this out a bit, with the target of some common setups on
production servers, in particular to allow specifying a version string
once and using it in a templated way. Some (quite minimal) examples in
the default config, but this can be used in most path fields as well as
pre_options, options, and the game binary name.