feat: make creeping frost damage fully resistable
As discussed in dev chat, this was not intended (a side effect of the
zap type), and has been a surprising damage spike for players. This
commit converts the damage part of this spell into BEAM_COLD so that it
is normally resisted, but otherwise keeps the damage the same. It also
documents flash freeze's resistance bypass (rime drakes etc), and
documents creeping frost's `frozen` status bypassing rC. I considered
whether to change rime drakes as well, but they are balanced in other
ways so it would need more consideration (breath timer, and flash freeze
itself does no damage with the frozen status already applied).