feat: templating enhancements for webtiles game config
Previously, webtiles allowed some minor string templating via %n and
(recently) %v, but that was it. However, in practice, on production
servers most game settings are repeated across many games, and some way
of editing them all it once is necessary to avoid errors. Most servers
(with only a few exceptions) dealt with this by running the
dgamelaunch-config package, which layered a bunch of string templating
on top of config.py (and other files).

This commit adds some relatively major enhancements to game definitions
in webtiles config that accomplishes the main goals here without needing
an external package. The main new feature here is the ability to specify
a partial game definition via a template, and then use that template to
fill out a full game definition. There's a bunch of ways to use this,
and it's probably somewhat overengineered even for the relatively
complex setups on current production servers, but my hope is that (a)
this will make life a lot easier for those managing large game configs,
and (b) this will make life a lot easier for anyone who wants to make
changes to how game configs are used. To give a flavor for how
simplified this is (see config.py and base.yaml for full examples), the
simplest use case is to define a template `default`, which will then be
used by default on any other templates and game definitions, that has
all the shared settings. The config.py example does this, allowing the
games definition itself to be written as just:

```
games = collections.OrderedDict([
    ("dcss-web-trunk",   dict(version = "trunk", name="Play %v",)),
    ("seeded-web-trunk", dict(version = "trunk", name="Seeded", options=["-seed"])),
    ("tut-web-trunk",    dict(version = "trunk", name="Tutorial", options=["-tutorial"])),
    ("sprint-web-trunk", dict(version = "trunk", name="Sprint %v", options=["-sprint"])),
])
```

This code may overall still need some bugs shaken out as it goes into
production, so server owners may prefer to let me test it out on CAO for
a while before rewriting their configs.

This commit along the way refactors a whole bunch of code.
1. Game loading is completely rewritten. It now supports game removals
on HUP, and treats config module games in exactly the same way that yaml
games are treated for loading purposes.
2. Game config string templating is restructured again: game configs are
now in a custom mapping class, so most code can just get templated
strings from objects of this class directly. (For convenience, I've
still kept the `game_param` function introduced in a recent commit.) A
bunch of older string templating code is cleaned up.
3. Lots of documentation. Though, this code now has a bit of an identity
crisis as to where this documentation should be; it's split between
config.py and base.yaml. Perhaps it should be in a third file.