fix: adjust `send_json_options` behavior when unset
This option was added in 0.14, and when unset was treated as False, I
assume for compatibility with versions that were around at the time.
However, very few servers even host versions this old now, and the code
also handles gracefully if the crawl binary errors when called with the
relevant option. I think it's time to get rid of this from active
webtiles configs, and so this commit accordinatly treats it as True if
it is unset. It can still be explicitly set to False for servers that
are still running relevant versions (this may be a good idea for
io-bound servers at least).