fix: adjust some game config validation details
Previously, config.py game configs weren't actually validated, so this
check seems to have some incorrect assumptions that flew under the
radar. `pre_options` was not declared at all, and `client_path` was
incorrectly described as required (if it is not set, the webserver will
pick it up from the crawl binary).