Don't crash when generating slings for ?acqs
There was a rare crash that could occur when reading a scroll of acquirement
where, when choosing a weapon, the now-defunct slings skill would be picked for
the kind of weapon that should be generated. Specifically,
_acquirement_weapon_subtype would attempt to iterate through every weapon trying
desperately to find something that uses SK_SLINGS before eventually dying out of
its main loop and returning 101 (OBJ_RANDOM), which was the initial value of
result. Later on, the value returned from _acquirement_weapon_subtype is blindly
used as the subtype for the object being generated, and /that/ subtype is then
used in an index for testing if the item is able to be wielded, and /that/
causes a segfault.