fix: cache some info across webserver reloads
Previously, `collect_game_modes` was rerun on a HUP. Unfortunately this
is a very heavy function with big game configs, as it runs the crawl
binary for each game and blocks the whole time. This fix caches the
gametype info for each binary across reloads, since in practice I think
this will never change; this function should now only block on initial
load (which is fine).