refactor: clean up save info collection
Previously, save info was collected for each game mode. In most
scenarios, this results in many unnecessary calls to crawl binaries,
because a single `-save-info-json USERNAME` call prints all saves for
that binary. Webtiles already had a mechanism for distinguishing
binaries that handles dgamelaunch-config setups correctly, so this
commit takes that and uses it to reduce the number of calls when
collecting save info for the lobby. Now, each binary is called at most
once. This commit also slightly cleans up how the callback chain is
constructing, and better comments this (slightly baffling) bit of code.