refactor: rewrite webtiles db code
Since the very beginning (82b49a24b7ed1), webtiles has accessed its user
databases by opening and then closing a connection on each access. I
haven't been able to figure out any reason why it should do things this
way (sqlite3 can certainly handle concurrency), and there are a bunch of
reasons why this is bad: (a) it increases overhead on each call (an
fopen is necessary each time, probably among other things), and (b) it
leads to a bunch of bad code to minimize the number of db opens per each
function.

This commit rewrites the db access code to have a persistent connection
to both of the sqlite3 databases. Because of this change, I was also
able to majorly refactor most of the db calls so that there is much less
duplicate code. This commit also changes (arguably) a bunch of stylistic
stuff in this code, and improves aspects of the API. Along the way, I
wrote unit tests for the major functions of this module.

I'm not sure how significant the speedup is, but it isn't nothing; on
a SSD in local testing this about halved the time of db calls (though
from ~20ms to ~10ms, so may or may not be impactful). But I suspect on
a server with a spinning disk (aka CAO), this will be quite noticeable.