feat: allow customizing recovery token lifetime
Also, default it to a longer value. Previous this was hardcoded to 1
hour, which is definitely a pretty safe option, but in practice was too
short for servers where admins are doing the password resets. Basically
it makes any admin-managed password reset that is not done synchronously
fail. In a quick survey of best practices here (or at least practices),
I found answers ranging from 1 to 24 hours. The previous value of 1 is
still probably practical for servers that have email password resets.