fix: more blocking annotation adjustments
Because the dgl file update runs on a timeout, it was getting false
positives for stuff that happens early in a connection, so explicitly
prevent the timeout version from doing this. Also, in my local testing,
some of these were false positives for MainHandler.get (namely the
initial render of client.html).