fix: add a unique index for dglusers.username
Despite being treated like it, username was neither indexed nor
constrainted to be unique. In large user dbs (aka CAO, with around 100k
rows), this was actually impacting performance -- lookup was on the
order of 100ms there. This commit creates new databases with a unique index,
and tries to add an index on upgrade. (Note that just a regular index
doesn't see noticeable performance gains without something else; I was
able to tune the queries in combination with a non-unique index, but I
think it's safe just to enforce uniqueness.)

I know that it's possible for this to fail, because uniqueness was not
previously enforced except by side effect (CAO has some duplicate rows
from 2007 right now), so it's handled relatively carefully. If an admin
encounters the error here on upgrade, they probably should go and
manually fix in their user db (typically you can do this by renaming one
of the duplicates, but you should check carefully).