fix: improve logging for RequestHandler and PeriodicCallback
For both of these, the slow callback logging printed fairly anonymous
info that made it hard to identify the actual source. This commit
improves the situations for both classes. For PeriodicCallback, I was
able to straightforwardly override `__repr__` since the object itself is
used. The method for RequestHandler is more entertaining than good.