feat: a webtiles config override specifically for live-debug mode
To test things on production servers in live-debug mode, I'm often
editing config.py in place. This isn't ideal (forgetting and not doing a
dgl publish could lead to these setting tweaks ending up in production).
So, add a way to override only settings for live-debug mode, based on
the config.yml override. Of course, to test a config as close to
possible as production, you probably don't want to heavily modify this.