Fix: Ref to scroll of ID may break in _identify()
Break caused by auto_assign_item_slot() and swap_inv_slots(),
when the ID'd item swaps with the scroll of ID.
Parameter 'link' is already provided to accommodate this.

Mainly, fixing this means scroll of ID will be assigned to its
assigned item_slot if it was previously unknown.

(Committer's note: simplified the check logic.)

Closes #2811.