Update the entrance message for the Pan rune shop vault
The other phrasing suggested to some that the pandemonium lord would
leave you alone if you paid up for the rune, which is not the case.
Now the announcement is a little more clear about how to get the rune,
without suggesting how the panlord thinks about the situation, one
way or another.

"nicolae why does the pan lord want to kill its customers"
i don't know buddy i'm not a pan lord