fix: don't allow treeform over deep water
80ed8d6641f added a check to allow transformations over dangerous
terrain when the player has tempflight, but this missed an interaction
with treeform. Treeform disables all sources of flight immediately, so
if the player had tempflight (e.g. via !flight), this would lead to
sequences like the following:

    You turn into a tree. Your roots penetrate the ground.
    You fall into the water!
    You can't survive in this terrain! You fly above the water, but the process is
    draining.
    Splash!
    You fall into the deep water. You sink to the bottom.

For the duration the player is in deep water, they get emergency flight
draining. (Permaflight had a correct version of this check.)  To fix,
move the tempflight check into `_flying_in_new_form` so the interaction
with forms forbidding flight is handled correctly.

Also, remove an obsolete Xom death check that looked at deep water/lava.