fix: don't keep processing webtiles ui-stack messages (dilly)
The `ui-stack` message gets sent when a spectator joins, and consists of
a sequence of menu descriptions. When a spectator was already viewing a
submenu, and a new spectator joined, the same menus would be added to
the old spectator's ui stack on top of other already displayed menus.
This commit simply prevents processing of this message more than once,
which is correct for the way it's currently used.