fix: don't update time display on spectator connection
This is a bit tricky to keep in sync, because the server-side version of
this works differently than the webtiles client version. (I'm inclined
to say the webtiles version is probably better and should be ported over
to the c++ side, but that's a bigger project.) This commit sends the
server-side time delta value on spectator connection, and uses the fact
that this is only sent by `_send_everything` to ignore time updates in
that case for the player. It's still possible for a spectator to be
seeing slightly out of sync values for this on connection, but it now
shouldn't ever be updated for the player on this case.