refactor: centralize some version-related metadata
The goal of this is to make it straightforward for forks to change this
metadata in one go, especially the bug report line (please send your bug
reports to you, not us). Minorly improve what info is displayed where,
e.g. show the build target with version number, show the bug tracker
link on the version page, a few other things.