feat: prevent good mutations from causing unfixable stat zero
Following 6a3c3ed53f1b it is possible to get stuck with stat zero via
only base stats + good mutations. For example, a starting feee with two
levels each of clever and agile has base 4 str, and -4 str from the
mutations. Recent example:
  https://crawl.xtahua.com/crawl/morgue/sanka/morgue-sanka-20221201-183043.txt
This char went from str 6 to 0 in one jiyva pass. (Note that the present
commit doesn't actually check mutation loss, but it would have prevented
the case in this morgue because loss happens before gain.)

This is very rare for normal players so I think just hasn't come up, but
on a jiyva char can happen rather easily (and rapidly) especially with
an early jiyva conversion.  This circumstance can effectively make a
char non-viable, with the only counterplays being mutation potion luck,
or preemptively raising stats.  A 4str char possibly *should* consider
the latter, but it doesn't really seem like a large set of "good"
mutations is an amazing way to reinforce this message, and a char like
the one above would have only ~1 opportunity to even do this.

The commit only checks muts flagged with mutflag::good, so
weak/dopey/clumsy could still cause stat 0 (not sure about whether they
should be able to, as they may have the same fixability issues, but the
circumstances are a lot different and would at least typically allow
more time to preemptively raise stats). The commit also doesn't check
items and so on, since if a mut causes stat 0 in combination with items,
there is an immediate counterplay. The mutation is silently rerolled if
it would cause stat 0.

To do this, I refactored how stats + mutations are handled to be a bit
more clearly data-ified. Interaction with the stats code is still
somewhat brute force, though.