fix: improve (with hacks) map sending to spectators
When a new spectator joins, `_send_everything` pushes out a whole bunch
of json data which previously was sent to all sockets connected to a
process. The biggest part of this is usually the map: a fully mapped
level can involve a 100k+ json, and handling this turns out to be
impactful, at least with many spectators. (I haven't profiled most
aspects of this, but what I do know is that the deflate time on cao can
be 3-10ms, which can get into human-noticeable territory with 10 or more
spectators. Even on a very fast local machine I can get this into what
are probably noticeable latencies with 20+ spectators.)

This commit adds some heuristics on the server side so that, after some
spectator(s) join, the next full map is sent only to those spectators.
This should take map handling from linear with the number of spectators,
to constant with the number of spectators. Webservers would need to
restart to get this change.

I also considered various more robust seeming versions of this,
including trying to compress the map message only once. But this turns
out to be extremely hard to do with how the deflate protocol works. And
I expect that there's also json decoding penalties for the map on the
receiver end. I may try to add something more systematic via control
messages, but this change also has the advantage that it should work
with any version of crawl that is currently played online.