fix: don't block with many spectators on game end
The `go_lobby` function is individually just a bit below the threshold
of being a problem (though it could use some testing in production), but
with a lot of spectators, this loop can easily be a problem when the
player saves or otherwise ends the game. Even in local testing I'm
seeing 500ms when 10+ spectators disconnect. This commit converts the
loop to one that is called asynchronously (for backwards compatibility,
via old-fashioned IOLoop callbacks).