feat: grey out useless items on the action panel
The game has a lot of situations when the player cannot use some
consumables. There are branch-, god-, and status effect-specific
item restrictions, some of which can be discovered only by trial
and error.

To better communicate these restrictions to players, grey out
currently useless items on the action panel.