fix: add more uselessness checks for consumables
There are several situations when the game doesn't let you use an item,
but the item is not marked as useless. For example, scrolls are not
considered useless in Gehenna or when you are confused or berserk.
Scrolls of teleportation are greyed out in the Gauntlet, but not when
you are under Stuck or -Tele status. Scrolls of enchant weapon/armour
look exactly the same in your inventory when you have something to
enchant and when there are no such items.

This commit adds several uselessness checks for consumables. This
affects not only the action panel, but also the `i`, `r`, `q`, and `v`
menus and the quiver.