feat: restful lobby endpoint
Currently lobby info can be extracted in one of two ways. First, like
beem, one can open a websockets connection and read the lobby json as it
comes in. Second, and lesser known, is the rube goldberg machine that
powers https://crawl.develz.org/. The way this works is that: (i) the
server writes a "dglwhere" file in a specific format, (ii) the webtiles
server admin exposes these via apache or the like[1], (iii) these are
collected and parsed by a python script[2] on cdo and turned into
aggregated json, and (iv) the cdo javascript side of things reads this
json from disk. Historically, this format presumably was used by servers
that were not running webtiles, but only one of those remains.

This commit puts forward something that I hope will eventually replace a
bunch of this, namely an endpoint that just directly generates the json
that scripts like this would need.

As usual, do not run bots on a server without permission, and don't
scrape this frequently enough that a server admin would notice.  This
endpoint is very lightweight, so it has no restrictions or throttling,
but it may need something if badly behaved bots are pointed at it (as
they seem to be for most crawl infrastructure, sooner or later).

[1] This is so lesser known that I think every server owner has
forgotten about this step, and they have faded away over time except for
cao. And it's not that the current cao admins actually remembered it
either.
[2] Script in question: https://github.com/crawl/dcss-website/blob/master/build-scripts/dgl-status-collect.py