build: comment and improve util/find_font
This is in response to:
https://github.com/crawl/crawl/pull/1367/commits/53545380446f6c79db6221757b9bb63b4259b878

That commit is wrong, because fc-match always tries to return something,
and if dejavu isn't installed, it will return something completely wrong
e.g. verdana. Also, it seems to have a false premise that this script
wasn't searching fontconfig directories -- because of the use of
fc-list, it does search whatever fontconfig directories are around (and
I've tested it out with macports' custom directories). However, this
wasn't very apparent on a casual read of the script, so I've added a
bunch of comments to clarify, and improved the fc-list call a bit.

The addition of "-L" here is from:
https://github.com/NixOS/nixpkgs/blob/master/pkgs/games/crawl/crawl_purify.patch
It seems useful + safe enough to include for everyone.  (That patch also
seems to have the false premise re fontconfig? So I've added comments to
head off the unnecessary change they included there.)