feat: show usability for scrolls/potions in description
The quality of scroll info here is a lot higher, because of the
existence of `cannot_read_item_reason`. Cancellation already had some
existing special casing, so I have included an informative message for
it, but no other potions will print a reason as of this commit.

To a lesser extent, this kind of info may also be useful for evokables,
and perhaps other items too.