fix: adjust more cases of perma vs temp uselessness
* Butterflies and summoning should be treated as perma-useless if allies
  are forbidden (either by oka or by ru).
* blinking/tele should be treated as perma-useless for formicid.

This also adjusts some of the scroll reason logic. I think this whole
code path could stand to be refactored so that the temp parameter is
propagated into the reason-generating code, and more perma-uselessness
reasons are provided. But for now, I have done some hacky brute force
checks to get this right, as well as reorded some of the reason
priorities.