Change the spelling of some words to Australian English.
Change how some words are spelt to reflect the standard spelling on
www.macquariedictionary.com.au.

This only changes things in places where the spellings may be displayed
to the user. I include the documentation in this. I haven't changed the
changelog, or any of the contrib stuff.

These include: armor => armour defense => defence offense => offence

I haven't changed the Longfellow quote, as it appears to use the poet's
original words.

Replace "alternate" with "alternative", as the dictionary describes the
"one possibility out of two" sense in which "alternate" is used as "a
disputed usage".