feat: Some tweaks to three lair vaults
pdpol_lair_cabin_in_the_woods:

Give the human survivor some passable gear. A good_item war axe and
lower tier good_item armour won't grant a lot loot value but will make
this human slightly less trivial to defeat.  Not to the extent that it's
likely to kill a player, but it should make fighting the survivor a
little more thematic.

Replace the lua block with a simple KMONS with a fallback definition.
Simplify some of the SUBST weights, and reglyph to put monsters on the
usual glyphs and terrain on glyphs not reserved for item slots.

pdpol_lair_cavern_descent:

Rebalance monster placement to reduce the overall monster count, since
it was pretty high for such a small and relatively open vault. Reglyph
to put monsters in rough order of difficulty and simplify statements.

For pdpol_lair_neglected_garden, a minor weight simplification.