fix: explicitly cast ghost colours to int
This prevents a warning in gcc. This probably means the related commit
in #2861 is unnecessary, though the sign issue still remains, as
detailed by the comment here. I spent a while trying to figure out how
to disable this warning, but when I looked into the details of this
exact bit of code, I think gcc was right to warn.