Adjust constriction
Remove the stepdown and scale linearly. Also, apply AC without breakpoints.
To compensate for increased damage after long durations, reduce the base
damage of a couple of the meatier constricting monsters, vile clutch, and
wand of roots. Damage is lower before 50 aut, so this may not have been
necessary, but we'll play it safe.

Closes #2883.