Add help feature to the options menu and some submenus.
If the user presses ? in an options menu, bring up the part of
options_guide.txt, arena.txt or fight_simulator.txt which describes
that option. The latter two are technically now available in the help
menu, but the player isn't told where they are.

Also add a help key to "choose from a list" options and when Crawl
displays a parsing error. There isn't a key for while editing a
"string" option.