Improve portability of the default menu colour settings.
Replace \b (which is not in the POSIX standard) with a space.

The menu colour prefix for an item always starts with "identified " or
"unidentified " (neither of which has a default colour by itself).
Option names do not contain a space. Use this to distinguish the two
without using a tag.