refactor: consolidate potion uselessness code
Previously, potion uselessness was handled in a superficial way by a
bunch of checks in item-use.cc, and then in a much more fine-grained way
on the actual quaff calls. The latter didn't expose the messaging to be
usable elsewhere (as scrolls now do), and wasn't fully consistent with
the other uselessness code. This commit refactors so that the potion
code overall uses the same strategy as scrolls, except that this is
plumbed directly into the `can_quaff` checks in potion.cc.

* Expand temp uselessness to match almost exactly when potions won't
  work. This does now mark e.g. !heal as temp useless at full hp and so
  on, which could be annoying. (nlavsky was worried about making the
  action panel too flickery with some of these cases.) If so, I'd rather
  have the general case working and then special case what ui needs
  further changes. A few weird cases: brilliance no longer works under
  Trog, before it worked but its effects only matter if you cased.
  !invis still works when temp useless with an override, and I've fixed
  some of these cases. It's possible that effects prevented by armour
  (robe of folly, robe of vines) should be perma-useless? But generally
  that's not how we handle equip.
* Show uselessness reasons in item descriptions.
* Fix some corner cases: invis while wielding eos, !magic with full mp,
  curing with full hp and nothing to cure, brilliance with folly (prev
  marked as useless but drinkable), some flight check issues
* Overall message consistency