feat: add a generalized unequip menu
This moves the take off and remove popups to use UseItemMenu, and
adds a general unequip pane to that menu. This new pane can be accessed
via CMD_UNEQUIP, bound right now to `c` as the only free letter. There
was various refactoring and tweaking in this commit to enable this, and
these actions are now triggerable via use_an_item with the appropriate
operatio value.

Aspirational notes:
* it might be better to connect the equip menus with the unequip menus
  somehow. Maybe allow unequipping from the equip menu? The easiest
  thing here would just be another hotkey that swaps between the two.
* It would be really great if error messages for equip/unequip would
  show in this menu somehow. Unfortunately, they are all directly
  printed all over the place via mpr, so this is a bit involved to get
  going.

Misc notes:
* I moved a lot of the error messaging onto MSGCH_PROMPT. I'm not sure
  if this or the regular channel is really better, but this was very
  inconsistent before. Seems likely to still be inconsistent.
* I marked jewellery as temp useless when in a form that doesn't allow
  it; this is for consistency with weapons/armour, which are already
  marked this way.