fix: handle vs and curing uselessness
Followup to 41eeaaa. The problem was not the ailment check, but that the
call to you.can_potion_heal with temp set to false would lead to !curing
being marked as perma-useless to Vs. This commit refactors the logic to
be clearer, since every species that can drink at all should be able to
use curing (as far as I can see).