fix: disable misc from acquirement
Reverts a change in 8e341f0f67. It seems that enabling misc also enables
wands, which was unintentional. ...and weird, it looks like that'd lead
to 1.5 weighting of wands in dungeon acquirement, and 0.5 of misc? see
_find_acquirement_subtype -- I suspect the random_choose there should be
removed. But I don't have time to investigate right now so I'm just
going to revert.