Make Doom Hounds' Howl Will-resistable
For consistency with other mental effects, and to make Grunn's Doom
Hounds a bit weaker, Doom Hounds' Howl can now be resisted with Will.
It's a max (200) power effect, so it has about a 85% chance of hitting
folks around 1 pip of will, 67% chance of hitting folks around 2 pips,
and 7% at 4 pips. Willpower is halved in Tar, of course, so good luck
with that... :)