Make Shadow Step and Blinkbolt check constriction
As with other movement-based abilities (Power Leap, Electric Charge,
regular blinks, etc), make using Shadow Step or Blinkbolt count as escape
attempts for constriction and cancel the ability on failure.

Storm Form is immune to constriction so this isn't actually relevant for
Blinkbolt in practice, but add the check anyway in case it becomes possible
to Blinkbolt in some other way.