New early unrand: Force Lance
The Force Lance is a Heavy lance (same weapon type as Wyrmbane -
a spear with +2 base dam) that has a 1/3 chance of knocking back
enemies it deals damage to, sending them 1-2 spaces back and
potentially causing extra collision. I've marked it as an early
unrand (like Throatcutter, etc) because I'm unsure if its damage
really works out for the late game, but it might be quite strong
for an early unrand. We can play with the numbers.

Yes, this is a pun. I'm not sorry.