Allow "Victory" consumables out of combat
Based on feedback from kate, this commit allows players to use consumables
while out of combat when wearing Victory. The intent here is to allow
use of strategic consumables like ?enchant armour or !experience when the
character is safe, while continuing to restrict all consumables in-combat.

I've implemented this as a check for hostile visible monsters, to head off
future tech debt re: maintaining a list of strategic consumables for this.

I've also removed the chance of stat gain from killing easy mons:
It was a low chance anyway, and it didn't quite fit the intended conduct.